Fix console error
